backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "SwasthWrap Backend API with FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bson>=0.5.10",
    "fastapi>=0.115.13",
    "httpx>=0.28.1",
    "python-jose[cryptography]>=3.3.0",
    "motor>=3.7.1",
    "openai>=1.90.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.0.0",
    "pydantic[email]>=2.11.7",
    "pymango>=0.1.1",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.3",
    "pymongo>=4.13.2",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "sarvamai>=0.1.7",
    "aiohttp>=3.12.13",
    "pydub>=0.25.1",
    "aiofiles>=24.1.0",
    "pypdf2>=3.0.1",
    "pdfplumber>=0.11.7",
    "python-docx>=1.2.0",
]
